pyproject.toml
[tool.poetry]
name = "poetry add django djangorestframework"
version = "0.1.0"
description = ""
authors = ["ron9295 <ron9295@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.14"
django = "^6.0.3"
djangorestframework = "^3.16.1"
gunicorn = "^25.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
